+++ src/frontend/src/pages/ProjectsPage/Projects.tsx
@@ -8,12 +8,14 @@ import { routes } from '../../utils/routes';
 import WBSDetails from '../WBSDetails';
 import CreateProjectForm from '../CreateProjectPage/CreateProjectForm';
 import ProjectsPage from './ProjectsPage';
+import CreateWorkPackageForm from '../WorkPackageForm/CreateWorkPackageForm';
 
 const Projects: React.FC = () => {
   return (
     <Switch>
       <Route path={routes.PROJECTS_OVERVIEW} component={ProjectsPage} />
       <Route path={routes.PROJECTS_ALL} component={ProjectsPage} />
+      <Route path={routes.WORK_PACKAGE_NEW} component={CreateWorkPackageForm} />
       <Route path={routes.PROJECTS_NEW} component={CreateProjectForm} />
       <Route path={routes.PROJECTS_BY_WBS} component={WBSDetails} />
       <Route path={routes.PROJECTS} component={ProjectsPage} />
