+++ src/frontend/src/pages/WorkPackageDetailPage/WorkPackagePage.tsx
@@ -9,17 +9,15 @@ import { useSingleWorkPackage } from '../../hooks/work-packages.hooks';
 import { useAuth } from '../../hooks/auth.hooks';
 import LoadingIndicator from '../../components/LoadingIndicator';
 import ErrorPage from '../ErrorPage';
-import WorkPackageEditContainer from './WorkPackageEditContainer/WorkPackageEditContainer';
 import WorkPackageViewContainer from './WorkPackageViewContainer/WorkPackageViewContainer';
 import { useQuery } from '../../hooks/utils.hooks';
-import { useHistory } from 'react-router-dom';
+import EditWorkPackageForm from '../WorkPackageForm/EditWorkPackageForm';
 
 interface WorkPackagePageProps {
   wbsNum: WbsNumber;
 }
 
 const WorkPackagePage: React.FC<WorkPackagePageProps> = ({ wbsNum }) => {
-  const history = useHistory();
   const query = useQuery();
   const { isLoading, isError, data, error } = useSingleWorkPackage(wbsNum);
   const [editMode, setEditMode] = useState<boolean>(query.get('edit') === 'true');
@@ -29,15 +27,7 @@ const WorkPackagePage: React.FC<WorkPackagePageProps> = ({ wbsNum }) => {
   if (isError) return <ErrorPage message={error?.message} />;
 
   if (editMode) {
-    return (
-      <WorkPackageEditContainer
-        workPackage={data!}
-        exitEditMode={() => {
-          setEditMode(false);
-          history.push(`${history.location.pathname}`);
-        }}
-      />
-    );
+    return <EditWorkPackageForm wbsNum={wbsNum} stateFunction={setEditMode} />;
   }
 
   return (
