+++ src/frontend/src/pages/WorkPackageForm/WorkPackageFormDetails.tsx
@@ -4,15 +4,14 @@
  */
 
 import { User, WorkPackageStage } from 'shared';
-import { fullNamePipe } from '../../../utils/pipes';
-import { FormControl, FormLabel, Grid, MenuItem, TextField, Typography } from '@mui/material';
-import ReactHookTextField from '../../../components/ReactHookTextField';
+import { FormControl, FormLabel, Grid, MenuItem, TextField } from '@mui/material';
 import { Control, Controller, FieldErrorsImpl } from 'react-hook-form';
 import { DatePicker } from '@mui/x-date-pickers';
-import NERAutocomplete from '../../../components/NERAutocomplete';
-import { WorkPackageEditFormPayload } from './WorkPackageEditContainer';
-import ChangeRequestDropdown from '../../../components/ChangeRequestDropdown';
-import { Box } from '@mui/system';
+import { WorkPackageFormViewPayload } from './WorkPackageFormView';
+import NERAutocomplete from '../../components/NERAutocomplete';
+import ReactHookTextField from '../../components/ReactHookTextField';
+import PageBlock from '../../layouts/PageBlock';
+import { fullNamePipe } from '../../utils/pipes';
 
 interface Props {
   lead?: string;
@@ -21,8 +20,8 @@ interface Props {
   setLead: (lead?: string) => void;
   usersForProjectLead: User[];
   usersForProjectManager: User[];
-  control: Control<WorkPackageEditFormPayload>;
-  errors: Partial<FieldErrorsImpl<WorkPackageEditFormPayload>>;
+  control: Control<WorkPackageFormViewPayload>;
+  errors: Partial<FieldErrorsImpl<WorkPackageFormViewPayload>>;
 }
 
 const WorkPackageFormDetails: React.FC<Props> = ({
@@ -46,7 +45,7 @@ const WorkPackageFormDetails: React.FC<Props> = ({
 
   const StageSelect = () => (
     <FormControl fullWidth>
-      <FormLabel>Work Package Stage</FormLabel>
+      <FormLabel>Stage Select</FormLabel>
       <Controller
         name="stage"
         control={control}
@@ -65,12 +64,9 @@ const WorkPackageFormDetails: React.FC<Props> = ({
   );
 
   return (
-    <Box>
-      <Typography variant="h5" sx={{ marginBottom: '10px', color: 'white' }}>
-        Project Details
-      </Typography>
-      <Grid container spacing={1} xs={12}>
-        <Grid item xs={12} md={4}>
+    <PageBlock title="Work Package Details">
+      <Grid container xs={12}>
+        <Grid item xs={12} md={5} sx={{ mt: 2, mr: 2 }}>
           <FormControl fullWidth>
             <FormLabel>Work Package Name</FormLabel>
             <ReactHookTextField
@@ -81,13 +77,7 @@ const WorkPackageFormDetails: React.FC<Props> = ({
             />
           </FormControl>
         </Grid>
-        <Grid item xs={12} md={3}>
-          <ChangeRequestDropdown control={control} name="crId" errors={errors} changeHeight={56} />
-        </Grid>
-        <Grid item xs={12} md={3}>
-          <StageSelect />
-        </Grid>
-        <Grid item xs={12} md={2}>
+        <Grid item xs={12} md={2} sx={{ mt: 2, mr: 2 }}>
           <FormControl fullWidth sx={{ overflow: 'hidden' }}>
             <FormLabel sx={{ whiteSpace: 'noWrap' }}>Start Date (YYYY-MM-DD)</FormLabel>
             <Controller
@@ -109,7 +99,10 @@ const WorkPackageFormDetails: React.FC<Props> = ({
             />
           </FormControl>
         </Grid>
-        <Grid item xs={12} md={2}>
+        <Grid item xs={12} md={2} sx={{ mt: 2, mr: 2 }}>
+          <StageSelect />
+        </Grid>
+        <Grid item xs={12} md={2} sx={{ mt: 2, mr: 2 }}>
           <FormControl fullWidth>
             <FormLabel>Duration</FormLabel>
             <ReactHookTextField
@@ -121,10 +114,10 @@ const WorkPackageFormDetails: React.FC<Props> = ({
             />
           </FormControl>
         </Grid>
-        <Grid item xs={12} md={5}>
+        <Grid item xs={12} md={6} sx={{ mt: 1 }}>
           <FormLabel> Project Lead</FormLabel>
           <NERAutocomplete
-            sx={{ width: '100%' }}
+            sx={{ mt: 1, width: '90%' }}
             id="project-lead-autocomplete"
             onChange={(_event, value) => setLead(value?.id)}
             options={usersForProjectLead.map(userToOption)}
@@ -133,10 +126,10 @@ const WorkPackageFormDetails: React.FC<Props> = ({
             value={userToOption(usersForProjectLead.find((user) => user.userId.toString() === lead))}
           />
         </Grid>
-        <Grid item xs={12} md={5}>
+        <Grid item xs={12} md={6} sx={{ mt: 1 }}>
           <FormLabel>Project Manager</FormLabel>
           <NERAutocomplete
-            sx={{ width: '100%' }}
+            sx={{ mt: 1, width: '90%' }}
             id="project-manager-autocomplete"
             onChange={(_event, value) => setManager(value?.id)}
             options={usersForProjectManager.map(userToOption)}
@@ -146,7 +139,7 @@ const WorkPackageFormDetails: React.FC<Props> = ({
           />
         </Grid>
       </Grid>
-    </Box>
+    </PageBlock>
   );
 };
