--- src/frontend/src/tests/pages/WorkPackageDetailPage/WorkPackageEditContainer/EditModeOptions.test.tsx
@@ -1,16 +1,0 @@
-/*
- * This file is part of NER's FinishLine and licensed under GNU AGPLv3.
- * See the LICENSE file in the repository root folder for details.
- */
-
-import { render, screen } from '../../../test-support/test-utils';
-import EditModeOptions from '../../../../pages/WorkPackageDetailPage/WorkPackageEditContainer/EditModeOptions';
-
-// TODO: better tests here
-describe('Renders edit mode options', () => {
-  test('renders all of the buttons', () => {
-    render(<EditModeOptions exitEditMode={() => {}} />);
-    expect(screen.getByText('Cancel')).toBeInTheDocument();
-    expect(screen.getByText('Save')).toBeInTheDocument();
-  });
-});
